:recycle: refactored file structure and other misc adjustments
