+++ server/bin/www
@@ -83,4 +83,5 @@ function onListening() {
     var addr = server.address();
     var bind = typeof addr === "string" ? "pipe " + addr : "port " + addr.port;
     debug("Listening on " + bind);
+    console.log("Server listening on: " + bind);
 }
