+++ server/tests/app.test.js
@@ -1,5 +1,5 @@
 const request = require("supertest");
-const app = require("./app");
+const app = require("../app");
 
 const origin = "http://localhost:3000"
