+++ server/tests/delete.todo.test.js
@@ -1,11 +1,9 @@
 const request = require("supertest");
-const app = require("./../../app");
-const database = require("./../../firebase").database;
+const app = require("../app");
+const database = require("../firebase").database;
+const getTodaysDate = require("../utils/date-helper");
 
-let NZGmt = 13;
-let todaysDate = new Date();
-todaysDate.setHours(todaysDate.getHours() + NZGmt);
-todaysDate = todaysDate.toISOString().slice(0, 10);
+let todaysDate = getTodaysDate()
 
 const expectedAfterDeleteTodo = [
     {
@@ -34,14 +32,14 @@ const expectedAfterDeleteTodo = [
     },
 ];
 
-jest.mock("./../../firebase", () => {
-    const mockDatabase = require("../../test_utils/mocks/mockDatabase");
+jest.mock("../firebase", () => {
+    const mockDatabase = require("../mocks/mockDatabase");
     return {
         database: mockDatabase,
     };
 });
 
-jest.mock("../../auth", () => {
+jest.mock("../utils/auth", () => {
     return jest.fn(() => Promise.resolve("mockValidToken"));
 });
