+++ server/tests/get.todo.test.js
@@ -1,10 +1,8 @@
 const request = require("supertest");
-const app = require("../../app");
+const app = require("../app");
+const getTodaysDate = require("../utils/date-helper");
 
-let NZGmt = 13;
-let todaysDate = new Date();
-todaysDate.setHours(todaysDate.getHours() + NZGmt);
-todaysDate = todaysDate.toISOString().slice(0, 10);
+let todaysDate = getTodaysDate()
 
 const expectedTodaysTodo = [
     {
@@ -73,14 +71,14 @@ const expectedAllTodo = [
     },
 ];
 
-jest.mock("./../../firebase", () => {
-    const mockDatabase = require("../../test_utils/mocks/mockDatabase");
+jest.mock("../firebase", () => {
+    const mockDatabase = require("../mocks/mockDatabase");
     return {
         database: mockDatabase,
     };
 });
 
-jest.mock("../../auth", () => {
+jest.mock("../utils/auth", () => {
     return jest.fn(() => Promise.resolve("mockValidToken"));
 });
